+++ R/data_rescale.R
@@ -120,6 +120,8 @@ rescale.grouped_df <- function(x,
                                to = c(0, 100),
                                range = NULL,
                                ignore_case = FALSE,
+                               regex = FALSE,
+                               verbose = FALSE,
                                ...) {
   info <- attributes(x)
 
@@ -127,7 +129,12 @@ rescale.grouped_df <- function(x,
   grps <- attr(x, "groups", exact = TRUE)
 
   # evaluate arguments
-  select <- .select_nse(select, x, exclude, ignore_case)
+  select <- .select_nse(select,
+                        x,
+                        exclude,
+                        ignore_case,
+                        regex = regex,
+                        verbose = verbose)
 
   # dplyr < 0.8.0?
   if (is.null(grps)) {
@@ -163,9 +170,16 @@ rescale.data.frame <- function(x,
                                to = c(0, 100),
                                range = NULL,
                                ignore_case = FALSE,
+                               regex = FALSE,
+                               verbose = FALSE,
                                ...) {
   # evaluate arguments
-  select <- .select_nse(select, x, exclude, ignore_case)
+  select <- .select_nse(select,
+                        x,
+                        exclude,
+                        ignore_case,
+                        regex = regex,
+                        verbose = verbose)
 
   # Transform the range so that it is a list now
   if (!is.null(range)) {